+++ design-system/Button.tsx
@@ -32,13 +32,13 @@ export const Button: React.FC<ButtonProps> = ({
   primary = false,
   size = 'medium',
   backgroundColor,
-  className = '',
+  className,
   label,
   ...props
 }) => {
   const BASE_BUTTON = 'uppercase tracking-wider box-border';
-  const PRIMARY_BUTTON = `${BASE_BUTTON} bg-blue-600 border border-blue-600 text-white focus:outline-none focus:bg-blue-800 focus:border-blue-800 hover:bg-blue-800 hover:border-blue-800`;
-  const SECONDARY_BUTTON = `${BASE_BUTTON} border border-gray-400 text-black focus:outline-none focus:border-gray-800 hover:border-gray-800`;
+  const PRIMARY_BUTTON = `${BASE_BUTTON} bg-primary-600 border border-primary-600 text-white focus:outline-none focus:bg-primary-900 focus:border-primary-900 hover:bg-primary-900 hover:border-primary-900`;
+  const SECONDARY_BUTTON = `${BASE_BUTTON} border bg-white text-primary-600 border-primary-600 text-black focus:outline-none focus:text-primary-900 focus:border-primary-900 hover:border-primary-900 hover:text-primary-900`;
 
   const SIZE_BUTTON = {
     small: 'rounded-sm text-sm px-2 py-1',