+++ CHANGELOG.md
@@ -6,16 +6,16 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-- graphQL ([#36](https://github.com/FabienGreard/front-end-rocks-boilerplate/issues/36)) ([3083241](https://github.com/FabienGreard/front-end-rocks-boilerplate/commit/30832414a87a786d6ba429e456432cd9c66c5318))
+- graphQL ([#36](https://github.com/FabienGreard/kirby-boilerplate/issues/36)) ([3083241](https://github.com/FabienGreard/kirby-boilerplate/commit/30832414a87a786d6ba429e456432cd9c66c5318))
 
 ## 1.2.0 (2021-03-22)
 
 ### Features
 
-- design-system ([#7](https://github.com/FabienGreard/front-end-rocks-boilerplate/issues/7)) ([cc4c825](https://github.com/FabienGreard/front-end-rocks-boilerplate/commit/cc4c8259c1afff4b60350a1a4b7eb7dec43cfdda))
+- design-system ([#7](https://github.com/FabienGreard/kirby-boilerplate/issues/7)) ([cc4c825](https://github.com/FabienGreard/kirby-boilerplate/commit/cc4c8259c1afff4b60350a1a4b7eb7dec43cfdda))
 
 ## 1.1.0 (2021-03-21)
 
 ### Features
 
-- pwa support ([#6](https://github.com/FabienGreard/front-end-rocks-boilerplate/issues/6)) ([d28489d](https://github.com/FabienGreard/front-end-rocks-boilerplate/commit/d28489d7d5b28bafb474bab38c6e87ee411c003b))
+- pwa support ([#6](https://github.com/FabienGreard/kirby-boilerplate/issues/6)) ([d28489d](https://github.com/FabienGreard/kirby-boilerplate/commit/d28489d7d5b28bafb474bab38c6e87ee411c003b))